Reverted to filepath as index
