+++ src/cleanvision/dataset/utils.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List, Optional, TYPE_CHECKING
 
-from cleanvision.dataset import Dataset
+from cleanvision.dataset.base_dataset import Dataset
 from cleanvision.dataset.folder_dataset import FolderDataset
 from cleanvision.dataset.hf_dataset import HFDataset
 from cleanvision.dataset.torch_dataset import TorchDataset
